requirements.txt
keras==2.14.0
numpy==1.26.1
Pillow==10.1.0
streamlit==1.28.1
